Update version 2.8
Added round number print and a guard to create timing analyser instance only once during the start of the game.

+++ ReactionGame.cydsn/source/asw/ReactionGame.h
@@ -109,8 +109,8 @@ typedef struct {
     RG_State_t rg_curState;  // Current analyzer state
     /* Measurement & Record Data */
     uint32_t rg_totalTime;   // total time taken for 10 rounds
-    uint32_t rg_score;       // score of the game
-    uint32_t rg_roundPlayed; // count of the rounds // Keep it here or local?
+    uint16_t rg_score;       // score of the game
+    uint16_t rg_roundPlayed; // count of the rounds
 } RG_t;
 
 // Wrapper to allow representing the file in Together as class
